+++ playwright.config.ts
@@ -1,4 +1,4 @@
-import type { PlaywrightTestConfig } from '@playwright/test';
+import { type PlaywrightTestConfig, devices } from '@playwright/test';
 
 const config: PlaywrightTestConfig = {
 	webServer: {
@@ -6,7 +6,29 @@ const config: PlaywrightTestConfig = {
 		port: 4173
 	},
 	testDir: 'tests',
-	testMatch: /(.+\.)?(test|spec)\.[jt]s/
+	testMatch: /(.+\.)?(test|spec)\.[jt]s/,
+	projects: [
+		{
+			name: 'chromium',
+			use: {
+				...devices['Desktop Chrome']
+			}
+		},
+
+		{
+			name: 'firefox',
+			use: {
+				...devices['Desktop Firefox']
+			}
+		},
+
+		{
+			name: 'webkit',
+			use: {
+				...devices['Desktop Safari']
+			}
+		}
+	]
 };
 
 export default config;