+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Dates should be in`YYYY-MM-DD` format.
 
 - Better tip styling.
 - Consistent currency link styling.
+- Better external link styling.
 
 ### Maintenance